Added pattern style examples, structured plugin
