+++ inc/css/pattern-styles.css
@@ -1,0 +1,7 @@
+* {
+    color: red !important;
+}
+
+.mbp-buttons {
+    border: 25px solid teal;
+}
